Fix compiler warnings for implicit int casts and dampen collisions. Also add vcpkg integration and CMakeList fixes

+++ src/physics/fixed_grid.cc
@@ -6,15 +6,15 @@ FixedGrid::FixedGrid(const Vec2f& top_left, const Vec2f& bottom_right)
     : top_left_{top_left}
     , bottom_right_{bottom_right}
 {
-    rows_ = ( bottom_right_.y - top_left_.y ) / CELL_SIZE;
-    cols_ = ( bottom_right_.x - top_left_.x ) / CELL_SIZE;
+    rows_ = static_cast<int>(( bottom_right_.y - top_left_.y ) / CELL_SIZE);
+    cols_ = static_cast<int>(( bottom_right_.x - top_left_.x ) / CELL_SIZE);
     grid_.resize(rows_ * cols_);
 }
 
 void FixedGrid::reset()
 {
-    rows_ = ( bottom_right_.y - top_left_.y ) / CELL_SIZE;
-    cols_ = ( bottom_right_.x - top_left_.x ) / CELL_SIZE;
+    rows_ = static_cast<int>(( bottom_right_.y - top_left_.y ) / CELL_SIZE);
+    cols_ = static_cast<int>(( bottom_right_.x - top_left_.x ) / CELL_SIZE);
     grid_.clear();
     grid_.resize(rows_ * cols_);
 }
@@ -22,8 +22,8 @@ void FixedGrid::reset()
 Vec2i FixedGrid::getCell(const Vec2f& position)
 {
     const Vec2f rel_pos = position - top_left_;
-    const uint8_t c = rel_pos.x / CELL_SIZE;
-    const uint8_t r = rel_pos.y / CELL_SIZE;
+    const int c = static_cast<int>(rel_pos.x / CELL_SIZE);
+    const int r = static_cast<int>(rel_pos.y / CELL_SIZE);
     return Vec2i{r, c};
 }
